plant-disease-api-main/api/requirements.txt
#tensorflow==2.13.0
fastapi
uvicorn
python-multipart
pillow
#tensorflow-serving-api==2.13.0
matplotlib
numpy
google-generativeai